Workaround https://github.com/scala/bug/issues/11803 - move all BIO* classes to package object
* trifunctorize BIOFork to BIOFork3
* workaround for https://github.com/scala/bug/issues/11803
