+++ fundamentals/fundamentals-bio/src/main/scala/izumi/functional/bio/BIOPrimitives.scala
@@ -2,7 +2,7 @@ package izumi.functional.bio
 
 import zio.{IO, Promise, Ref, Semaphore}
 
-trait BIOPrimitives[F[_, _]] extends BIOPrimitivesInstances {
+trait BIOPrimitives[F[+_, +_]] extends BIOPrimitivesInstances {
   def mkRef[A](a: A): F[Nothing, BIORef[F, A]]
   def mkLatch[A]: F[Nothing, BIOPromise[F, Nothing, A]]
   def mkPromise[E, A]: F[Nothing, BIOPromise[F, E, A]]
