+++ fundamentals/fundamentals-bio/src/test/scala/izumi/fundamentals/bio/test/BIOSyntaxTest.scala
@@ -7,10 +7,13 @@ import org.scalatest.WordSpec
 import scala.concurrent.duration._
 
 object masking {
-  import izumi.functional.bio.{BIOFork, BIOPrimitives}
+  import izumi.functional.bio.{BIOFork, BIOFork3, BIOPrimitives}
 
-  type Primitives[F[_, _]] = BIOPrimitives[F]
-  type Fork[F[_, _]] = BIOFork[F]
+  type Primitives[F[+_, +_]] = BIOPrimitives[F]
+  type Primitives3[F[-_, +_, +_]] = BIOPrimitives[F[Any, +?, +?]]
+  type Fork[F[+_, +_]] = BIOFork[F]
+  type Fork3[F[-_, +_, +_]] = BIOFork3[F]
+  type BIOMonad3[F[-_, +_, +_]] = BIOMonad[F[Any, +?, +?]]
 }
 
 class BIOSyntaxTest extends WordSpec {
@@ -50,15 +53,20 @@ class BIOSyntaxTest extends WordSpec {
       F.map(z[F])(_ => ())
     }
     def `attach BIOPrimitives & BIOFork methods even when they aren't imported`[F[+_, +_]: BIOMonad: Primitives: Fork]: F[Nothing, Int] = {
-      F.fork[Nothing, Int] {
+      F.fork[Any, Nothing, Int] {
         F.mkRef(4).flatMap(r => r.update(_ + 5) *> r.get.map(_ - 1))
       }.flatMap(_.join)
     }
-    lazy val a = x
-    lazy val b = y[zio.IO](_: BIOAsync[zio.IO])
-    lazy val c = z
-    lazy val d = `attach BIOPrimitives & BIOFork methods even when they aren't imported`[zio.IO]
-    lazy val _ = (a, b, c, d)
+    def `attach BIOPrimitives & BIOFork3 methods to a trifunctor BIO even when not imported`[F[-_, +_, +_]: BIOMonad3: Primitives3: Fork3]: F[Any, Nothing, Int] = {
+      F.fork(F.mkRef(4).flatMap(r => r.update(_ + 5) *> r.get.map(_ - 1))).flatMap(_.join)
+    }
+    lazy val _ = (
+      x,
+      y[zio.IO](_: BIOAsync[zio.IO]),
+      z,
+      `attach BIOPrimitives & BIOFork methods even when they aren't imported`[zio.IO],
+      `attach BIOPrimitives & BIOFork3 methods to a trifunctor BIO even when not imported`[zio.ZIO],
+    )
   }
 
 }
